Rename project/package to remove the Attributed suffix
We're no longer just about attributes, conventions play a role too now :)
