+++ src/Attributed/readme.md
@@ -1,0 +1,4 @@
+# Obsolete
+
+This package now just references [Devlooped.Extensions.DependencyInjection](http://nuget.org/packages/Devlooped.Extensions.DependencyInjection) 
+which contains all its functionality and more.
