--- src/DependencyInjection.Attributed.Tests/ContentFiles.targets
@@ -1,16 +1,0 @@
-<Project>
-  <!-- Simulates including the files under contentFiles/cs/netstandard2.0 in the final package -->
-
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)..\DependencyInjection.Attributed\AttributedServicesExtension.cs" Link="AttributedServicesExtension.cs" Visible="false" />
-    <Compile Include="$(MSBuildThisFileDirectory)..\DependencyInjection.Attributed\ServiceAttribute.cs" Link="ServiceAttribute.cs" Visible="false" />
-    <Compile Include="$(MSBuildThisFileDirectory)..\DependencyInjection.Attributed\ServiceAttribute`1.cs" Link="ServiceAttribute`1.cs" Visible="false" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)..\DependencyInjection.Attributed\AttributedServicesExtension.cs" Type="Non-Resx" />
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)..\DependencyInjection.Attributed\ServiceAttribute.cs" Type="Non-Resx" />
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)..\DependencyInjection.Attributed\ServiceAttribute`1.cs" Type="Non-Resx" />
-  </ItemGroup>
-
-</Project>
