+++ src/DependencyInjection.Tests/ComponentModelTests.cs
