+++ src/DependencyInjection.Tests/CompositionTests.cs
