+++ src/DependencyInjection.Tests/ConventionsTests.cs
