+++ src/DependencyInjection.Tests/DependencyInjection.Tests.csproj
@@ -1,24 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <Import Project="..\DependencyInjection.Attributed\Devlooped.Extensions.DependencyInjection.Attributed.props" />
+  <Import Project="..\DependencyInjection\Devlooped.Extensions.DependencyInjection.props" />
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
     <EmitCompilerGeneratedFiles>true</EmitCompilerGeneratedFiles>
     <RootNamespace>Tests</RootNamespace>
   </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Remove="ComponentModelTests.cs" />
-    <Compile Remove="CompositionTests.cs" />
-    <Compile Remove="GenerationTests.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="ComponentModelTests.cs" />
-    <None Include="CompositionTests.cs" />
-    <None Include="GenerationTests.cs" />
-  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.9.0" />
@@ -32,7 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\DependencyInjection.Attributed\Attributed.csproj" ReferenceOutputAssembly="false" OutputItemType="Analyzer" />
+    <ProjectReference Include="..\DependencyInjection\DependencyInjection.csproj" ReferenceOutputAssembly="false" OutputItemType="Analyzer" />
   </ItemGroup>
 
   <ItemGroup>
@@ -41,6 +29,6 @@
   </ItemGroup>
 
   <Import Project="ContentFiles.targets" />
-  <Import Project="..\DependencyInjection.Attributed\Devlooped.Extensions.DependencyInjection.Attributed.targets" />
+  <Import Project="..\DependencyInjection\Devlooped.Extensions.DependencyInjection.targets" />
 
 </Project>
