+++ src/DependencyInjection.Tests/GenerationTests.cs
