+++ src/DependencyInjection/SymbolExtensions.cs
